requirements.txt
rich>=13.0.0
typer>=0.9.0
python-dotenv>=1.0.0
gitpython>=3.1.0
pydantic>=2.0.0
click>=8.0.0
streamlit>=1.28.0
streamlit-ace>=0.1.1
watchdog>=3.0.0